Converted GangMember to tsx, only Gang left
