+++ src/Augmentation/ui/PurchasedAugmentations.tsx
@@ -7,7 +7,6 @@ import * as React from "react";
 import { Augmentations } from "../../Augmentation/Augmentations";
 import { AugmentationNames } from "../../Augmentation/data/AugmentationNames";
 import { Player } from "../../Player";
-import { IPlayerOwnedAugmentation } from "../../Augmentation/PlayerOwnedAugmentation";
 
 import { AugmentationAccordion } from "../../ui/React/AugmentationAccordion";
